backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
neo4j==5.14.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
anthropic==0.7.7
python-multipart==0.0.6

# 认证相关
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# 其他依赖
python-dotenv==1.0.0
slowapi==0.1.9
cryptography>=41.0.0
